feat(dag): rebase, use waterfall for put
