+++ src/dag/index.js
@@ -1,0 +1,12 @@
+'use strict'
+
+const moduleConfig = require('../utils/module-config')
+
+module.exports = (arg) => {
+  const send = moduleConfig(arg)
+
+  return {
+    get: require('./get')(send),
+    put: require('./put')(send)
+  }
+}
